Move ambient type declarations to namespaces
This confers three benefits:
* All type declarations are now emitted to dist as intended
* This project's types won't conflict with `@types/screeps`, etc
* Consumers can still access this project's types from the global
  namespace via the `Api` namespace
